Move aliases over to common and share it with unicodedata.
