+++ compiler/parser/Cargo.toml
@@ -20,6 +20,7 @@ tiny-keccak = { version = "2", features = ["sha3"] }
 [dependencies]
 rustpython-ast = { path = "../ast", version = "0.2.0" }
 rustpython-compiler-core = { path = "../core", version = "0.2.0" }
+rustpython-common = { path = "../../common" }
 
 ahash = { workspace = true }
 itertools = { workspace = true }
